Workaround for auth flow getting stuck
This is a dumb workaround for an issue that I hope to track down and
fix properly in the future.

The issue is that sometimes tud_hid_set_report_cb is called with
report_id=0xF0, but the contents of the buffer are clearly not the
auth payload, but wheel force feedback data instead.

This commit makes it so that the current auth flow iteration still
fails, because we're missing parts of the nonce, but at least it's not
getting stuck and the PS5 is giving us an 8 minute grace period so if
one of the next iterations succeeds we're okay.